--- src/main/java/base/PetType.java
@@ -1,4 +1,0 @@
-package base;
-
-public enum PetType {
-}
