--- src/main/java/entity/Pet.java
@@ -1,4 +1,0 @@
-package entity;
-
-public class Pet {
-}
