--- src/main/java/entity/Request.java
@@ -1,4 +1,0 @@
-package entity;
-
-public class Request {
-}
