--- src/main/java/entity/Response.java
@@ -1,4 +1,0 @@
-package entity;
-
-public class Response {
-}
