--- src/main/java/entity/Review.java
@@ -1,4 +1,0 @@
-package entity;
-
-public class Review {
-}
