--- src/main/java/entity/User.java
@@ -1,4 +1,0 @@
-package entity;
-
-public class User {
-}
