+++ src/main/java/org/rami/petis/base/PetType.java
@@ -1,0 +1,4 @@
+package org.rami.petis.base;
+
+public enum PetType {
+}
