+++ src/main/java/org/rami/petis/base/RequestStatus.java
@@ -1,4 +1,4 @@
-package base;
+package org.rami.petis.base;
 
 public enum RequestStatus {
 }
