+++ src/main/java/org/rami/petis/base/ReviewGrade.java
@@ -1,4 +1,4 @@
-package base;
+package org.rami.petis.base;
 
 public enum ReviewGrade {
 }
