+++ src/main/java/org/rami/petis/base/UserType.java
@@ -1,4 +1,4 @@
-package base;
+package org.rami.petis.base;
 
 public enum UserType {
     ADMIN(0),OWNER(1),SITTER(2),BOTH(3);
