+++ src/main/java/org/rami/petis/entity/Request.java
@@ -1,0 +1,4 @@
+package org.rami.petis.entity;
+
+public class Request {
+}
