+++ src/main/java/org/rami/petis/entity/User.java
@@ -1,0 +1,4 @@
+package org.rami.petis.entity;
+
+public class User {
+}
